Fixed issue such that only the owner of the image can delete the image

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -27,6 +27,8 @@ public class ImageController {
     @Autowired
     private TagService tagService;
 
+    private String error = "Only the owner of the image can edit the image";
+
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
     public String getUserImages(Model model) {
@@ -92,13 +94,22 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        Integer userId = image.getUser().getId();
+        User user = (User) session.getAttribute("loggeduser");
+        Integer loggedInUserId = user.getId();
+        //Only the owner of the image can edit the image
+        if(!userId.equals(loggedInUserId)) {
+            model.addAttribute("editError", error);
+            model.addAttribute("image", image);
+        } else {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        return "images/image";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +151,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        Integer userId = image.getUser().getId();
+        User user = (User) session.getAttribute("loggeduser");
+        Integer loggedInUserId = user.getId();
+        //Only the owner of the image can edit the image
+        if(!userId.equals(loggedInUserId)) {
+            model.addAttribute("deleteError", error);
+            model.addAttribute("image", image);
+        } else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        return "images/image";
     }
